Split the flag-gated sync-to-remote state handlers per context instead of branching on `is_called_from_{traverse,code_hosting}`
The `_handle_untracked_branch`, `_handle_untracked_state`, `_handle_ahead_state`, `_handle_diverged_and_newer_state` and `__pick_remote` helpers in `base.py`
used to take `is_called_from_traverse` / `is_called_from_code_hosting` flags and behave differently depending on them,
which was confusing and error-prone even though it shared some code.

Give each call context its own flag-free copy instead: the code-hosting variants live in `MacheteClientWithCodeHosting`,
the traverse variants in `TraverseMacheteClient`, and `_handle_behind_state` / `_handle_diverged_and_older_state`
(never gated, only ever used by traverse) move out of `base.py` into `TraverseMacheteClient` as well.

This is a behavior-preserving refactor with no test changes, with one nuance: the old shared `_handle_ahead_state`,
when reached from the code-hosting flow, accepted the un-advertised `yq` answer at its `(y, N, q)` prompt as a plain push without quitting
(the `yq`-quit was suppressed by `is_called_from_traverse`), unlike the `diverged-and-newer` handler which does quit on `yq`.
The code-hosting `ahead` handler now treats `yq` as a non-affirmative answer, exactly like any other unoffered input at that prompt;
this also keeps the `enforce-yq-check-for-each-y-yes-yq-check` invariant satisfied once the methods are no longer shared.
