+++ git_machete/client/base.py
@@ -3,19 +3,24 @@ import shlex
 import sys
 import textwrap
 from enum import Enum, auto
-from typing import Callable, Dict, Iterator, List, NoReturn, Optional, Sequence, Tuple, TypeVar
+from typing import (Callable, Dict, Iterator, List, NoReturn, Optional,
+                    Sequence, Tuple, TypeVar)
 
 from git_machete.client import branch_layout
 from git_machete.client.state import MacheteState, ManagedBranchName
 from git_machete.config import MacheteConfig, SquashMergeDetection
-from git_machete.constants import INITIAL_COMMIT_COUNT_FOR_LOG, TOTAL_COMMIT_COUNT_FOR_LOG
-from git_machete.git import (HEAD, AnyBranchName, AnyRevision, BranchPair, ForkPointOverrideData, FullCommitHash, Git, LocalBranchShortName,
-                             RemoteBranchShortName, SyncToRemoteStatus)
+from git_machete.constants import (INITIAL_COMMIT_COUNT_FOR_LOG,
+                                   TOTAL_COMMIT_COUNT_FOR_LOG)
+from git_machete.git import (HEAD, AnyBranchName, AnyRevision, BranchPair,
+                             ForkPointOverrideData, FullCommitHash, Git,
+                             LocalBranchShortName, RemoteBranchShortName,
+                             SyncToRemoteStatus)
 from git_machete.utils import fs
 from git_machete.utils.cmd import run_cmd
 from git_machete.utils.collections import excluding, get_second, tupled
 from git_machete.utils.debug_log import debug
-from git_machete.utils.exceptions import InteractionStopped, MacheteException, UnexpectedMacheteException
+from git_machete.utils.exceptions import (MacheteException,
+                                          UnexpectedMacheteException)
 from git_machete.utils.markup import input_fmt, pretty_choices, print_fmt, warn
 from git_machete.utils.paths import AbsPath, Path
 
@@ -842,293 +847,6 @@ class MacheteClient:
                                         cwd=self._git.get_current_worktree_root_dir())
             if exit_code != 0:
                 raise MacheteException(f"The machete-post-slide-out hook exited with {exit_code}, aborting.")
-
-    # === Sync-to-remote state handlers ===
-
-    def __pick_remote(
-            self,
-            *,
-            branch: LocalBranchShortName,
-            is_called_from_traverse: bool,
-            is_called_from_code_hosting: bool,
-            opt_push_untracked: bool,
-            opt_push_tracked: bool,
-            opt_yes: bool
-    ) -> None:
-        rems = self._git.get_remotes()
-        print("\n".join(f"[{index + 1}] {rem}" for index, rem in enumerate(rems)))
-        if is_called_from_traverse:
-            msg = f"Select number 1..{len(rems)} to specify the destination remote " \
-                "repository, or 'n' to skip this branch, or " \
-                "'q' to quit the traverse: "
-        else:
-            msg = f"Select number 1..{len(rems)} to specify the destination remote " \
-                "repository, or 'q' to quit the operation: "
-
-        ans = input(msg).lower()
-        if ans in ('q', 'quit'):
-            raise InteractionStopped
-        try:
-            index = int(ans) - 1
-            if index not in range(len(rems)):
-                raise MacheteException(f"Invalid index: {index + 1}")
-            self._handle_untracked_branch(
-                new_remote=rems[index],
-                branch=branch,
-                is_called_from_traverse=is_called_from_traverse,
-                is_called_from_code_hosting=is_called_from_code_hosting,
-                opt_push_untracked=opt_push_untracked,
-                opt_push_tracked=opt_push_tracked,
-                opt_yes=opt_yes)
-        except ValueError:
-            if is_called_from_code_hosting:
-                raise MacheteException('Could not establish remote repository, operation interrupted.')
-
-    def _handle_untracked_branch(
-            self,
-            *,
-            new_remote: str,
-            branch: LocalBranchShortName,
-            is_called_from_traverse: bool,
-            is_called_from_code_hosting: bool,
-            opt_push_untracked: bool,
-            opt_push_tracked: bool,
-            opt_yes: bool
-    ) -> None:
-        remotes: List[str] = self._git.get_remotes()
-        can_pick_other_remote = len(remotes) > 1 and is_called_from_traverse
-        other_remote_choice = "o[ther-remote]" if can_pick_other_remote else ""
-        remote_branch = RemoteBranchShortName.of(f"{new_remote}/{branch}")
-        if not self._git.get_commit_hash_by_revision(remote_branch):
-            choices = pretty_choices(
-                *('y', 'N', 'q', 'yq', other_remote_choice) if is_called_from_traverse else ('y', 'Q'))
-            ask_message = f"Push untracked branch <b>{branch}</b> to <b>{new_remote}</b>?" + choices
-            ask_opt_yes_message = f"Pushing untracked branch <b>{branch}</b> to <b>{new_remote}</b>..."
-            ans = self.ask_if(
-                ask_message,
-                ask_opt_yes_message,
-                opt_yes=opt_yes,
-                override_answer=None if opt_push_untracked else "N")
-            if is_called_from_traverse:
-                if ans in ('y', 'yes', 'yq'):
-                    self._git.push(new_remote, branch)
-                    if ans == 'yq':
-                        raise InteractionStopped
-                elif can_pick_other_remote and ans in ('o', 'other'):
-                    self.__pick_remote(
-                        branch=branch,
-                        is_called_from_traverse=is_called_from_traverse,
-                        is_called_from_code_hosting=is_called_from_code_hosting,
-                        opt_push_untracked=opt_push_untracked,
-                        opt_push_tracked=opt_push_tracked,
-                        opt_yes=opt_yes)
-                elif ans in ('q', 'quit'):
-                    raise InteractionStopped
-                return
-            else:
-                if ans in ('y', 'yes'):
-                    self._git.push(new_remote, branch)
-                else:
-                    raise InteractionStopped
-                return
-
-        relation = self._git.get_relation_to_remote_counterpart(branch, remote_branch)
-
-        message: str = {
-            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE:
-                f"Branch <b>{branch}</b> is untracked, but its remote counterpart candidate <b>{remote_branch}</b> "
-                f"already exists and both branches point to the same commit.",
-            SyncToRemoteStatus.BEHIND_REMOTE:
-                f"Branch <b>{branch}</b> is untracked, but its remote counterpart candidate <b>{remote_branch}</b> "
-                f"already exists and is ahead of <b>{branch}</b>.",
-            SyncToRemoteStatus.AHEAD_OF_REMOTE:
-                f"Branch <b>{branch}</b> is untracked, but its remote counterpart candidate <b>{remote_branch}</b> "
-                f"already exists and is behind <b>{branch}</b>.",
-            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE:
-                f"Branch <b>{branch}</b> is untracked, it diverged from its remote counterpart candidate <b>{remote_branch}</b>, "
-                f"and has <b>older</b> commits than <b>{remote_branch}</b>.",
-            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE:
-                f"Branch <b>{branch}</b> is untracked, it diverged from its remote counterpart candidate <b>{remote_branch}</b>, "
-                f"and has <b>newer</b> commits than <b>{remote_branch}</b>."
-        }[SyncToRemoteStatus(relation)]
-
-        ask_message, ask_opt_yes_message = {
-            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE: (
-                f"Set the remote of <b>{branch}</b> to <b>{new_remote}</b> without pushing or pulling?" +
-                pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
-                f"Setting the remote of <b>{branch}</b> to <b>{new_remote}</b>..."
-            ),
-            SyncToRemoteStatus.BEHIND_REMOTE: (
-                f"Pull <b>{branch}</b> (fast-forward only) from <b>{new_remote}</b>?" +
-                pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
-                f"Pulling <b>{branch}</b> (fast-forward only) from <b>{new_remote}</b>..."
-            ),
-            SyncToRemoteStatus.AHEAD_OF_REMOTE: (
-                f"Push branch <b>{branch}</b> to <b>{new_remote}</b>?" + pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
-                f"Pushing branch <b>{branch}</b> to <b>{new_remote}</b>..."
-            ),
-            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE: (
-                f"Reset branch <b>{branch}</b> to the commit pointed by <b>{remote_branch}</b>?" +
-                pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
-                f"Resetting branch <b>{branch}</b> to the commit pointed by <b>{remote_branch}</b>..."
-            ),
-            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE: (
-                f"Push branch <b>{branch}</b> with force-with-lease to <b>{new_remote}</b>?" +
-                pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
-                f"Pushing branch <b>{branch}</b> with force-with-lease to <b>{new_remote}</b>..."
-            )
-        }[SyncToRemoteStatus(relation)]
-
-        override_answer: Optional[str] = {
-            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE: None,
-            SyncToRemoteStatus.BEHIND_REMOTE: None,
-            SyncToRemoteStatus.AHEAD_OF_REMOTE: None if opt_push_tracked else "N",
-            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE: None,
-            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE: None if opt_push_tracked else "N",
-        }[SyncToRemoteStatus(relation)]
-
-        yes_action: Callable[[], None] = {
-            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE: lambda: self._git.set_upstream_to(remote_branch),
-            SyncToRemoteStatus.BEHIND_REMOTE: lambda: self._git.pull_ff_only(new_remote, remote_branch),
-            SyncToRemoteStatus.AHEAD_OF_REMOTE: lambda: self._git.push(new_remote, branch),
-            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE: lambda: self._git.reset_keep(remote_branch),
-            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE: lambda: self._git.push(
-                new_remote, branch, force_with_lease=True)
-        }[SyncToRemoteStatus(relation)]
-
-        print_fmt(message)
-        ans = self.ask_if(ask_message, ask_opt_yes_message, override_answer=override_answer, opt_yes=opt_yes)
-        if ans in ('y', 'yes', 'yq'):
-            yes_action()
-            if ans == 'yq':
-                raise InteractionStopped
-        elif ans in ('q', 'quit'):
-            raise InteractionStopped
-
-    def _handle_untracked_state(
-            self,
-            *,
-            branch: LocalBranchShortName,
-            is_called_from_traverse: bool,
-            is_called_from_code_hosting: bool,
-            opt_push_tracked: bool,
-            opt_push_untracked: bool,
-            opt_yes: bool
-    ) -> None:
-        remotes: List[str] = self._git.get_remotes()
-        self._ensure_blank_separator()
-        if len(remotes) == 1:
-            self._handle_untracked_branch(
-                new_remote=remotes[0],
-                branch=branch,
-                is_called_from_traverse=is_called_from_traverse,
-                is_called_from_code_hosting=is_called_from_code_hosting,
-                opt_push_untracked=opt_push_untracked,
-                opt_push_tracked=opt_push_tracked,
-                opt_yes=opt_yes)
-        elif "origin" in remotes:
-            self._handle_untracked_branch(
-                new_remote="origin",
-                branch=branch,
-                is_called_from_traverse=is_called_from_traverse,
-                is_called_from_code_hosting=is_called_from_code_hosting,
-                opt_push_untracked=opt_push_untracked,
-                opt_push_tracked=opt_push_tracked,
-                opt_yes=opt_yes)
-        else:
-            # We know that there is at least 1 remote, otherwise sync-to-remote state would be NO_REMOTES and not UNTRACKED
-            print_fmt(f"Branch <b>{branch}</b> is untracked and there's no <b>origin</b> remote.")
-            self.__pick_remote(
-                branch=branch,
-                is_called_from_traverse=is_called_from_traverse,
-                is_called_from_code_hosting=is_called_from_code_hosting,
-                opt_push_untracked=opt_push_untracked,
-                opt_push_tracked=opt_push_tracked,
-                opt_yes=opt_yes)
-
-    def _handle_ahead_state(
-            self,
-            *,
-            current_branch: LocalBranchShortName,
-            remote: str,
-            is_called_from_traverse: bool,
-            opt_push_tracked: bool,
-            opt_yes: bool
-    ) -> None:
-        self._ensure_blank_separator()
-        choices = pretty_choices(*('y', 'N', 'q', 'yq') if is_called_from_traverse else ('y', 'N', 'q'))
-        ans = self.ask_if(
-            f"Push <b>{current_branch}</b> to <b>{remote}</b>?" + choices,
-            f"Pushing <b>{current_branch}</b> to <b>{remote}</b>...",
-            override_answer=None if opt_push_tracked else "N",
-            opt_yes=opt_yes
-        )
-        if ans in ('y', 'yes', 'yq'):
-            self._git.push(remote, current_branch)
-            if ans == 'yq' and is_called_from_traverse:
-                raise InteractionStopped
-        elif ans in ('q', 'quit'):
-            raise InteractionStopped
-
-    def _handle_behind_state(self, *, branch: LocalBranchShortName, remote: str, opt_yes: bool) -> None:
-        self._ensure_blank_separator()
-        remote_branch = self._git.get_combined_counterpart_for_fetching_of_branch(branch)
-        assert remote_branch is not None
-        ans = self.ask_if(
-            f"Branch <b>{branch}</b> is behind its remote counterpart <b>{remote_branch}</b>.\n"
-            f"Pull <b>{branch}</b> (fast-forward only) from <b>{remote}</b>?" + pretty_choices('y', 'N', 'q', 'yq'),
-            f"Branch <b>{branch}</b> is behind its remote counterpart <b>{remote_branch}</b>.\n"
-            f"Pulling <b>{branch}</b> (fast-forward only) from <b>{remote}</b>...",
-            opt_yes=opt_yes)
-        if ans in ('y', 'yes', 'yq'):
-            self._git.pull_ff_only(remote, remote_branch)
-            if ans == 'yq':
-                raise InteractionStopped
-        elif ans in ('q', 'quit'):
-            raise InteractionStopped
-
-    def _handle_diverged_and_newer_state(
-            self,
-            *,
-            current_branch: LocalBranchShortName,
-            remote: str,
-            is_called_from_traverse: bool = True,
-            opt_push_tracked: bool,
-            opt_yes: bool
-    ) -> None:
-        self._ensure_blank_separator()
-        remote_branch = self._git.get_combined_counterpart_for_fetching_of_branch(current_branch)
-        assert remote_branch is not None
-        choices = pretty_choices(*('y', 'N', 'q', 'yq') if is_called_from_traverse else ('y', 'N', 'q'))
-        ans = self.ask_if(
-            f"Branch <b>{current_branch}</b> diverged from (and has newer commits than) its remote counterpart <b>{remote_branch}</b>.\n"
-            f"Push <b>{current_branch}</b> with force-with-lease to <b>{remote}</b>?" + choices,
-            f"Branch <b>{current_branch}</b> diverged from (and has newer commits than) its remote counterpart <b>{remote_branch}</b>.\n"
-            f"Pushing <b>{current_branch}</b> with force-with-lease to <b>{remote}</b>...",
-            override_answer=None if opt_push_tracked else "N", opt_yes=opt_yes)
-        if ans in ('y', 'yes', 'yq'):
-            self._git.push(remote, current_branch, force_with_lease=True)
-            if ans == 'yq':
-                raise InteractionStopped
-        elif ans in ('q', 'quit'):
-            raise InteractionStopped
-
-    def _handle_diverged_and_older_state(self, branch: LocalBranchShortName, *, opt_yes: bool) -> None:
-        self._ensure_blank_separator()
-        remote_branch = self._git.get_combined_counterpart_for_fetching_of_branch(branch)
-        assert remote_branch is not None
-        ans = self.ask_if(
-            f"Branch <b>{branch}</b> diverged from (and has older commits than) its remote counterpart <b>{remote_branch}</b>.\n"
-            f"Reset branch <b>{branch}</b> to the commit pointed by <b>{remote_branch}</b>?" + pretty_choices('y', 'N', 'q', 'yq'),
-            f"Branch <b>{branch}</b> diverged from (and has older commits than) its remote counterpart <b>{remote_branch}</b>.\n"
-            f"Resetting branch <b>{branch}</b> to the commit pointed by <b>{remote_branch}</b>...",
-            opt_yes=opt_yes)
-        if ans in ('y', 'yes', 'yq'):
-            self._git.reset_keep(remote_branch)
-            if ans == 'yq':
-                raise InteractionStopped
-        elif ans in ('q', 'quit'):
-            raise InteractionStopped
 
     # === Output formatting ===
