+++ git_machete/client/with_code_hosting.py
@@ -1,20 +1,26 @@
 import itertools
 import os
-from typing import Dict, Iterator, List, Optional, Set, Tuple
+from typing import Callable, Dict, Iterator, List, Optional, Set, Tuple
 
 from git_machete.annotation import Annotation, Qualifiers
 from git_machete.client.state import ManagedBranchName
 from git_machete.client.status import StatusMacheteClient
-from git_machete.code_hosting import (CodeHostingApi, CodeHostingSpec, OrganizationAndRepository, OrganizationAndRepositoryAndRemote,
+from git_machete.code_hosting import (CodeHostingApi, CodeHostingSpec,
+                                      OrganizationAndRepository,
+                                      OrganizationAndRepositoryAndRemote,
                                       PullRequest, is_matching_remote_url)
 from git_machete.config import PRDescriptionIntroStyle, SquashMergeDetection
-from git_machete.git import GitFormatPatterns, GitLogEntry, LocalBranchShortName, RemoteBranchShortName, SyncToRemoteStatus
+from git_machete.git import (GitFormatPatterns, GitLogEntry,
+                             LocalBranchShortName, RemoteBranchShortName,
+                             SyncToRemoteStatus)
 from git_machete.utils import date
 from git_machete.utils.collections import find_or_none, get_non_empty_lines
 from git_machete.utils.debug_log import debug
-from git_machete.utils.exceptions import MacheteException, UnexpectedMacheteException
+from git_machete.utils.exceptions import (InteractionStopped, MacheteException,
+                                          UnexpectedMacheteException)
 from git_machete.utils.fs import slurp_file
-from git_machete.utils.markup import colored_yes_no, green_ok, pretty_choices, print_fmt, warn
+from git_machete.utils.markup import (colored_yes_no, green_ok, pretty_choices,
+                                      print_fmt, warn)
 from git_machete.utils.paths import AbsPath
 
 
@@ -163,26 +169,22 @@ class MacheteClientWithCodeHosting(StatusMacheteClient):
             if anno is None or anno.qualifiers.push:
                 if s == SyncToRemoteStatus.AHEAD_OF_REMOTE:
                     assert remote is not None
-                    self._handle_ahead_state(
+                    self.__handle_ahead_state(
                         current_branch=current_branch,
                         remote=remote,
-                        is_called_from_traverse=False,
                         opt_push_tracked=True,
                         opt_yes=opt_yes)
                 elif s == SyncToRemoteStatus.UNTRACKED:
-                    self._handle_untracked_state(
+                    self.__handle_untracked_state(
                         branch=current_branch,
-                        is_called_from_traverse=False,
-                        is_called_from_code_hosting=True,
                         opt_push_tracked=True,
                         opt_push_untracked=True,
                         opt_yes=opt_yes)
                 elif s == SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE:
                     assert remote is not None
-                    self._handle_diverged_and_newer_state(
+                    self.__handle_diverged_and_newer_state(
                         current_branch=current_branch,
                         remote=remote,
-                        is_called_from_traverse=False,
                         opt_push_tracked=True,
                         opt_yes=opt_yes)
                 else:
@@ -319,11 +321,9 @@ class MacheteClientWithCodeHosting(StatusMacheteClient):
 
         # Handle missing base branch (push it if necessary)
         if not base_branch_found_on_remote:
-            self._handle_untracked_branch(
+            self.__handle_untracked_branch(
                 branch=base,
                 new_remote=base_org_repo_remote.remote,
-                is_called_from_traverse=False,
-                is_called_from_code_hosting=True,
                 opt_push_tracked=False,
                 opt_push_untracked=True,
                 opt_yes=opt_yes)
@@ -462,26 +462,22 @@ class MacheteClientWithCodeHosting(StatusMacheteClient):
 
             if s == SyncToRemoteStatus.AHEAD_OF_REMOTE:
                 assert remote is not None
-                self._handle_ahead_state(
+                self.__handle_ahead_state(
                     current_branch=current_branch,
                     remote=remote,
-                    is_called_from_traverse=False,
                     opt_push_tracked=True,
                     opt_yes=True)
             elif s == SyncToRemoteStatus.UNTRACKED:
-                self._handle_untracked_state(
+                self.__handle_untracked_state(
                     branch=current_branch,
-                    is_called_from_traverse=False,
-                    is_called_from_code_hosting=True,
                     opt_push_tracked=True,
                     opt_push_untracked=True,
                     opt_yes=True)
             elif s == SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE:
                 assert remote is not None
-                self._handle_diverged_and_newer_state(
+                self.__handle_diverged_and_newer_state(
                     current_branch=current_branch,
                     remote=remote,
-                    is_called_from_traverse=False,
                     opt_push_tracked=True,
                     opt_yes=True)
             else:
@@ -991,3 +987,212 @@ class MacheteClientWithCodeHosting(StatusMacheteClient):
             return result
 
         raise UnexpectedMacheteException("All params passed to __get_applicable_pull_requests are empty.")
+
+    # === Sync-to-remote state handlers (code-hosting variants) ===
+    #
+    # These intentionally duplicate the shape of the traverse-side handlers (see TraverseMacheteClient)
+    # rather than sharing a single flag-gated implementation:
+    # the code-hosting flow never offers the `yq` (push and quit) / other-remote choices nor swallows an unparsable
+    # remote selection, so keeping the two flows apart is clearer than threading `is_called_from_*` flags through `base.py`.
+
+    def __handle_untracked_branch(
+            self,
+            *,
+            new_remote: str,
+            branch: LocalBranchShortName,
+            opt_push_untracked: bool,
+            opt_push_tracked: bool,
+            opt_yes: bool
+    ) -> None:
+        remote_branch = RemoteBranchShortName.of(f"{new_remote}/{branch}")
+        if not self._git.get_commit_hash_by_revision(remote_branch):
+            ask_message = f"Push untracked branch <b>{branch}</b> to <b>{new_remote}</b>?" + pretty_choices('y', 'Q')
+            ask_opt_yes_message = f"Pushing untracked branch <b>{branch}</b> to <b>{new_remote}</b>..."
+            ans = self.ask_if(
+                ask_message,
+                ask_opt_yes_message,
+                opt_yes=opt_yes,
+                override_answer=None if opt_push_untracked else "N")
+            if ans in ('y', 'yes'):
+                self._git.push(new_remote, branch)
+            else:
+                raise InteractionStopped
+            return
+
+        relation = self._git.get_relation_to_remote_counterpart(branch, remote_branch)
+
+        message: str = {
+            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE:
+                f"Branch <b>{branch}</b> is untracked, but its remote counterpart candidate <b>{remote_branch}</b> "
+                f"already exists and both branches point to the same commit.",
+            SyncToRemoteStatus.BEHIND_REMOTE:
+                f"Branch <b>{branch}</b> is untracked, but its remote counterpart candidate <b>{remote_branch}</b> "
+                f"already exists and is ahead of <b>{branch}</b>.",
+            SyncToRemoteStatus.AHEAD_OF_REMOTE:
+                f"Branch <b>{branch}</b> is untracked, but its remote counterpart candidate <b>{remote_branch}</b> "
+                f"already exists and is behind <b>{branch}</b>.",
+            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE:
+                f"Branch <b>{branch}</b> is untracked, it diverged from its remote counterpart candidate <b>{remote_branch}</b>, "
+                f"and has <b>older</b> commits than <b>{remote_branch}</b>.",
+            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE:
+                f"Branch <b>{branch}</b> is untracked, it diverged from its remote counterpart candidate <b>{remote_branch}</b>, "
+                f"and has <b>newer</b> commits than <b>{remote_branch}</b>."
+        }[SyncToRemoteStatus(relation)]
+
+        ask_message, ask_opt_yes_message = {
+            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE: (
+                f"Set the remote of <b>{branch}</b> to <b>{new_remote}</b> without pushing or pulling?" +
+                pretty_choices('y', 'N', 'q', 'yq'),
+                f"Setting the remote of <b>{branch}</b> to <b>{new_remote}</b>..."
+            ),
+            SyncToRemoteStatus.BEHIND_REMOTE: (
+                f"Pull <b>{branch}</b> (fast-forward only) from <b>{new_remote}</b>?" +
+                pretty_choices('y', 'N', 'q', 'yq'),
+                f"Pulling <b>{branch}</b> (fast-forward only) from <b>{new_remote}</b>..."
+            ),
+            SyncToRemoteStatus.AHEAD_OF_REMOTE: (
+                f"Push branch <b>{branch}</b> to <b>{new_remote}</b>?" + pretty_choices('y', 'N', 'q', 'yq'),
+                f"Pushing branch <b>{branch}</b> to <b>{new_remote}</b>..."
+            ),
+            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE: (
+                f"Reset branch <b>{branch}</b> to the commit pointed by <b>{remote_branch}</b>?" +
+                pretty_choices('y', 'N', 'q', 'yq'),
+                f"Resetting branch <b>{branch}</b> to the commit pointed by <b>{remote_branch}</b>..."
+            ),
+            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE: (
+                f"Push branch <b>{branch}</b> with force-with-lease to <b>{new_remote}</b>?" +
+                pretty_choices('y', 'N', 'q', 'yq'),
+                f"Pushing branch <b>{branch}</b> with force-with-lease to <b>{new_remote}</b>..."
+            )
+        }[SyncToRemoteStatus(relation)]
+
+        override_answer: Optional[str] = {
+            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE: None,
+            SyncToRemoteStatus.BEHIND_REMOTE: None,
+            SyncToRemoteStatus.AHEAD_OF_REMOTE: None if opt_push_tracked else "N",
+            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE: None,
+            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE: None if opt_push_tracked else "N",
+        }[SyncToRemoteStatus(relation)]
+
+        yes_action: Callable[[], None] = {
+            SyncToRemoteStatus.IN_SYNC_WITH_REMOTE: lambda: self._git.set_upstream_to(remote_branch),
+            SyncToRemoteStatus.BEHIND_REMOTE: lambda: self._git.pull_ff_only(new_remote, remote_branch),
+            SyncToRemoteStatus.AHEAD_OF_REMOTE: lambda: self._git.push(new_remote, branch),
+            SyncToRemoteStatus.DIVERGED_FROM_AND_OLDER_THAN_REMOTE: lambda: self._git.reset_keep(remote_branch),
+            SyncToRemoteStatus.DIVERGED_FROM_AND_NEWER_THAN_REMOTE: lambda: self._git.push(
+                new_remote, branch, force_with_lease=True)
+        }[SyncToRemoteStatus(relation)]
+
+        print_fmt(message)
+        ans = self.ask_if(ask_message, ask_opt_yes_message, override_answer=override_answer, opt_yes=opt_yes)
+        if ans in ('y', 'yes', 'yq'):
+            yes_action()
+            if ans == 'yq':
+                raise InteractionStopped
+        elif ans in ('q', 'quit'):
+            raise InteractionStopped
+
+    def __handle_untracked_state(
+            self,
+            *,
+            branch: LocalBranchShortName,
+            opt_push_tracked: bool,
+            opt_push_untracked: bool,
+            opt_yes: bool
+    ) -> None:
+        remotes: List[str] = self._git.get_remotes()
+        self._ensure_blank_separator()
+        if len(remotes) == 1:
+            self.__handle_untracked_branch(
+                new_remote=remotes[0],
+                branch=branch,
+                opt_push_untracked=opt_push_untracked,
+                opt_push_tracked=opt_push_tracked,
+                opt_yes=opt_yes)
+        elif "origin" in remotes:
+            self.__handle_untracked_branch(
+                new_remote="origin",
+                branch=branch,
+                opt_push_untracked=opt_push_untracked,
+                opt_push_tracked=opt_push_tracked,
+                opt_yes=opt_yes)
+        else:
+            # We know that there is at least 1 remote, otherwise sync-to-remote state would be NO_REMOTES and not UNTRACKED
+            print_fmt(f"Branch <b>{branch}</b> is untracked and there's no <b>origin</b> remote.")
+            self.__pick_remote(
+                branch=branch,
+                opt_push_untracked=opt_push_untracked,
+                opt_push_tracked=opt_push_tracked,
+                opt_yes=opt_yes)
+
+    def __pick_remote(
+            self,
+            *,
+            branch: LocalBranchShortName,
+            opt_push_untracked: bool,
+            opt_push_tracked: bool,
+            opt_yes: bool
+    ) -> None:
+        rems = self._git.get_remotes()
+        print("\n".join(f"[{index + 1}] {rem}" for index, rem in enumerate(rems)))
+        msg = f"Select number 1..{len(rems)} to specify the destination remote " \
+            "repository, or 'q' to quit the operation: "
+        ans = input(msg).lower()
+        if ans in ('q', 'quit'):
+            raise InteractionStopped
+        try:
+            index = int(ans) - 1
+            if index not in range(len(rems)):
+                raise MacheteException(f"Invalid index: {index + 1}")
+            self.__handle_untracked_branch(
+                new_remote=rems[index],
+                branch=branch,
+                opt_push_untracked=opt_push_untracked,
+                opt_push_tracked=opt_push_tracked,
+                opt_yes=opt_yes)
+        except ValueError:
+            raise MacheteException('Could not establish remote repository, operation interrupted.')
+
+    def __handle_ahead_state(
+            self,
+            *,
+            current_branch: LocalBranchShortName,
+            remote: str,
+            opt_push_tracked: bool,
+            opt_yes: bool
+    ) -> None:
+        self._ensure_blank_separator()
+        ans = self.ask_if(
+            f"Push <b>{current_branch}</b> to <b>{remote}</b>?" + pretty_choices('y', 'N', 'q'),
+            f"Pushing <b>{current_branch}</b> to <b>{remote}</b>...",
+            override_answer=None if opt_push_tracked else "N",
+            opt_yes=opt_yes
+        )
+        if ans in ('y', 'yes'):
+            self._git.push(remote, current_branch)
+        elif ans in ('q', 'quit'):
+            raise InteractionStopped
+
+    def __handle_diverged_and_newer_state(
+            self,
+            *,
+            current_branch: LocalBranchShortName,
+            remote: str,
+            opt_push_tracked: bool,
+            opt_yes: bool
+    ) -> None:
+        self._ensure_blank_separator()
+        remote_branch = self._git.get_combined_counterpart_for_fetching_of_branch(current_branch)
+        assert remote_branch is not None
+        ans = self.ask_if(
+            f"Branch <b>{current_branch}</b> diverged from (and has newer commits than) its remote counterpart <b>{remote_branch}</b>.\n"
+            f"Push <b>{current_branch}</b> with force-with-lease to <b>{remote}</b>?" + pretty_choices('y', 'N', 'q'),
+            f"Branch <b>{current_branch}</b> diverged from (and has newer commits than) its remote counterpart <b>{remote_branch}</b>.\n"
+            f"Pushing <b>{current_branch}</b> with force-with-lease to <b>{remote}</b>...",
+            override_answer=None if opt_push_tracked else "N", opt_yes=opt_yes)
+        if ans in ('y', 'yes', 'yq'):
+            self._git.push(remote, current_branch, force_with_lease=True)
+            if ans == 'yq':
+                raise InteractionStopped
+        elif ans in ('q', 'quit'):
+            raise InteractionStopped
